+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [NEXT] - YYYY-MM-DD
+## [0.0.8] - 2021-05-05
 
 - Add config for unplaceable items
 - Add safer tile entity getting for non-essential methods